- fix missing cast for 32bit

+++ libs/SalesforceSDKCore/SalesforceSDKCore/Classes/SmartStore/SFAlterSoupLongOperation.m
@@ -81,10 +81,10 @@
 - (NSString*) description
 {
     return [NSString stringWithFormat:@"AlterSoupOperation = {rowId=%lld soupName=%@ soupTableName=%@ afterStep=%lu reIndexData=%@ oldIndexSpecs=%@ newIndexSpecs=%@}\n",
-            self.rowId,
+            (long long)self.rowId,
             self.soupName,
             self.soupTableName,
-            self.afterStep,
+            (unsigned long)self.afterStep,
             self.reIndexData ? @"YES" : @"NO",
             [SFJsonUtils JSONRepresentation:[SFSoupIndex asArrayOfDictionaries:self.oldIndexSpecs withColumnName:YES]],
             [SFJsonUtils JSONRepresentation:[SFSoupIndex asArrayOfDictionaries:self.indexSpecs  withColumnName:YES]]
